feat(metadata): ajout de métadonnées pour les pages principales et création d'un composant ButtonBack
