+++ src/app/(base-technique)/layout.tsx
@@ -1,4 +1,10 @@
 import Header from '@/components/Header';
+import { Metadata } from 'next';
+
+export const metadata: Metadata = {
+  title: "Axignis - Base technique de références",
+};
+
 export default function RootLayout({
   children,
 }: Readonly<{
