+++ src/app/(main-site)/(auth)/demande-devis-souscription/page.tsx
@@ -1,5 +1,10 @@
 import { Plans } from '@/types/plans';
+import { Metadata } from 'next';
 import RegisterForm from './RegisterForm';
+
+export const metadata: Metadata = {
+  title: "Axignis - Demande de devis de souscription",
+};
 
 export default async function RegisterPage({
   searchParams,
