+++ src/app/(main-site)/page.tsx
@@ -1,8 +1,13 @@
 import CtaContact from '@/components/CtaContact';
 import ServiceCard from '@/components/ServiceCard';
+import { Metadata } from 'next';
 import { getTranslations } from 'next-intl/server';
 import Image from 'next/image';
 import Link from 'next/link';
+
+export const metadata: Metadata = {
+  title: "Axignis - Accueil",
+};
 
 export default async function Home() {
   const t = await getTranslations();
